e-media1/pyproject.toml
[tool.poetry]
name = "e-media1"
version = "0.1.0"
description = ""
authors = ["jkusnierz01 <jedrek33@onet.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.9.0.80"
pillow = "^10.3.0"
matplotlib = "^3.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
